moved package code to src/NextCloud/; tests insert src/NextCloud to sys.path; renamed NextCloud js param to json_output
