+++ .travis.yml
@@ -7,7 +7,7 @@ before_script:
 
 # commands to run tests
 script:
-  - docker-compose run --rm python-api pytest
+  - docker-compose run --rm python-api python -m pytest
   - docker-compose run --rm python-api python example.py
 
 after_script:
