+++ src/NextCloud/__init__.py
@@ -1,0 +1,1 @@
+from .NextCloud import NextCloud
