+++ tests/__init__.py
@@ -1,0 +1,5 @@
+import sys
+from os.path import dirname
+from os.path import join
+
+sys.path.insert(0, join(dirname(dirname(__file__)), 'src'))
